+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change the default value of `band-resolved` from `on` (1) to `off` (0) in `castepkit-shg`.
 - Add `LD_LIBRARY_PATH` env to config.
+- Add `ome.py` for `calculate_ome_impi_XTIPC` and `atom_cutting.py` for `atom_cutting_impi_XTIPC`.
 
 
 ## [0.0.3] - 2025-05-27